+++ datamodel/odata-client-apache-httpclient5/src/main/java/com/sap/cloud/sdk/datamodel/odata/client/query/Order.java
@@ -1,0 +1,19 @@
+package com.sap.cloud.sdk.datamodel.odata.client.query;
+
+/**
+ * Used with <b>orderBy</b> methods in entity fluent helper objects to set the sorting order of field values.
+ *
+ * @since 5.31.0
+ */
+public enum Order
+{
+    /**
+     * Sort field values in ascending order.
+     */
+    ASC,
+
+    /**
+     * Sort field values in descending order.
+     */
+    DESC
+}