+++ datamodel/odata-client-apache-httpclient5/src/main/java/com/sap/cloud/sdk/datamodel/odata/client/request/ODataRequestResultMultipart.java
@@ -1,0 +1,21 @@
+package com.sap.cloud.sdk.datamodel.odata.client.request;
+
+import javax.annotation.Nonnull;
+
+/**
+ * Generic type of an OData request result.
+ *
+ * @since 5.31.0
+ */
+public interface ODataRequestResultMultipart
+{
+    /**
+     * Get the result from the OData batch response for a specific sub-request.
+     *
+     * @param request
+     *            The request to look for in the OData batch response.
+     * @return The OData result, that was extracted from the original OData batch response.
+     */
+    @Nonnull
+    ODataRequestResult getResult( @Nonnull final ODataRequestGeneric request );
+}